# src/states-districts/dto/state.dto.ts
import { HttpStatus } from '@nestjs/common'
import { Field, Int, ObjectType } from '@nestjs/graphql'
import { ApiProperty } from '@nestjs/swagger'

@ObjectType()
export class State {
  @ApiProperty({ example: '60d9910be37d6d6ed18bec11' })
  @Field()
  id: string

  @ApiProperty({ example: '16' })
  @Field(() => Int)
  state_id: number

  @ApiProperty({ example: 'Karnataka' })
  @Field()
  state_name: string
}

@ObjectType()
export class States {
  @ApiProperty()
  @Field()
  status: HttpStatus

  @ApiProperty({ type: State, isArray: true })
  @Field(() => [State])
  data: State[]
}