# swagger-spec.json
{"openapi":"3.0.0","info":{"title":"Covid Warrior","description":"List of all Covid Warrior APIs","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"State":{"type":"object","properties":{"id":{"type":"string","example":"60d9910be37d6d6ed18bec11"},"state_id":{"type":"number","example":"16"},"state_name":{"type":"string","example":"Karnataka"}},"required":["id","state_id","state_name"]},"States":{"type":"object","properties":{"status":{"type":"number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/State"}}},"required":["status","data"]},"District":{"type":"object","properties":{"district_id":{"type":"number"},"district_name":{"type":"string"}},"required":["district_id","district_name"]},"Districts":{"type":"object","properties":{"status":{"type":"number"},"id":{"type":"string"},"state_id":{"type":"number"},"districts":{"type":"array","items":{"$ref":"#/components/schemas/District"}}},"required":["status","id","state_id","districts"]},"LoginInput":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"UserDetails":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"username":{"type":"string"}},"required":["id","name","username"]},"LoginUserResponse":{"type":"object","properties":{"userDetails":{"$ref":"#/components/schemas/UserDetails"},"token":{"type":"string"}},"required":["userDetails","token"]},"RegisterInput":{"type":"object","properties":{"name":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}},"required":["name","username","password"]},"RegisterResponse":{"type":"object","properties":{"response":{"type":"string"},"statusCode":{"type":"number"}},"required":["response","statusCode"]},"VaccinationCenter":{"type":"object","properties":{"center_id":{"type":"number"},"name":{"type":"string"},"address":{"type":"string"},"state_name":{"type":"string"},"district_name":{"type":"string"},"block_name":{"type":"string"},"pincode":{"type":"number"},"from":{"type":"string"},"to":{"type":"string"},"lat":{"type":"number"},"long":{"type":"number"},"fee_type":{"type":"string"},"session_id":{"type":"string"},"date":{"type":"string"},"available_capacity":{"type":"number"},"available_capacity_dose1":{"type":"number"},"available_capacity_dose2":{"type":"number"},"fee":{"type":"string"},"allow_all_age":{"type":"boolean"},"min_age_limit":{"type":"number"},"vaccine":{"type":"string"},"slots":{"type":"array","items":{"type":"string"}}},"required":["center_id","name","address","state_name","district_name","block_name","pincode","from","to","lat","long","fee_type","session_id","date","available_capacity","available_capacity_dose1","available_capacity_dose2","fee","allow_all_age","min_age_limit","vaccine","slots"]},"VaccinationCenters":{"type":"object","properties":{"status":{"type":"number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/VaccinationCenter"}}},"required":["status","data"]},"CreateBeneficiary":{"type":"object","properties":{"name":{"type":"string"},"birth_year":{"type":"string","example":"DD-MM-YYYY"},"gender_id":{"type":"string"},"photo_id_type":{"type":"string"},"photo_id_number":{"type":"string"},"comorbidity_ind":{"type":"string","default":"N"},"consent_version":{"type":"string"}},"required":["name","birth_year","gender_id","photo_id_type","photo_id_number","comorbidity_ind","consent_version"]},"BeneficiaryResponse":{"type":"object","properties":{"status":{"type":"number"},"message":{"type":"string"}},"required":["status","message"]},"Beneficiary":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"birth_year":{"type":"string"},"gender_id":{"type":"number"},"photo_id_type":{"type":"number"},"photo_id_number":{"type":"string"},"comorbidity_ind":{"type":"string"},"consent_version":{"type":"string"},"scheduled":{"type":"boolean"}},"required":["id","name","birth_year","gender_id","photo_id_type","photo_id_number","comorbidity_ind","consent_version","scheduled"]},"Beneficiaries":{"type":"object","properties":{"status":{"type":"number"},"beneficiaries":{"type":"array","items":{"$ref":"#/components/schemas/Beneficiary"}}},"required":["status","beneficiaries"]},"CreateSchedule":{"type":"object","properties":{"beneficiaryId":{"type":"string"},"centerID":{"type":"string"},"slot":{"type":"string"},"date":{"type":"string"},"vaccine":{"type":"string"}},"required":["beneficiaryId","centerID","slot","date","vaccine"]},"ScheduleResponse":{"type":"object","properties":{"status":{"type":"number"},"message":{"type":"string"}},"required":["status","message"]},"Schedule":{"type":"object","properties":{"beneficiaryId":{"type":"string"},"centerID":{"type":"string"},"slot":{"type":"string"},"date":{"type":"string"},"vaccine":{"type":"string"}},"required":["beneficiaryId","centerID","slot","date","vaccine"]},"UpdateSchedule":{"type":"object","properties":{"centerID":{"type":"string"},"slot":{"type":"string"},"date":{"type":"string"},"vaccine":{"type":"string"}},"required":["centerID","slot","date","vaccine"]}}},"paths":{"/states-districts/states":{"get":{"operationId":"StatesDistrictsController_getStatesList","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/States"}}}}},"tags":["State and Districts APIs"]}},"/states-districts/districts/{stateId}":{"get":{"operationId":"StatesDistrictsController_getDistrictsList","parameters":[{"name":"stateId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Districts"}}}},"404":{"description":"State Not Found"}},"tags":["State and Districts APIs"]}},"/auth/login":{"post":{"operationId":"UserController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserResponse"}}}},"401":{"description":"Invalid credentials"}},"tags":["Authentication APIs"]}},"/auth/register":{"post":{"operationId":"UserController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"409":{"description":"User name already exists"},"501":{"description":"Invalid password pattern"}},"tags":["Authentication APIs"]}},"/vaccinationCenters/findVaccinationCenter":{"get":{"operationId":"VaccinationCenterController_findCenters","parameters":[{"name":"date","required":true,"in":"query","schema":{"type":"string"}},{"name":"pincode","required":false,"in":"query","schema":{"type":"number"}},{"name":"district_name","required":false,"in":"query","schema":{"type":"string"}},{"name":"latitude","required":false,"in":"query","schema":{"type":"number"}},{"name":"longitude","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VaccinationCenters"}}}}},"tags":["Vaccination Center APIs"]}},"/beneficiary":{"post":{"operationId":"BeneficiaryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBeneficiary"}}}},"responses":{"201":{"description":"Beneficiary added","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BeneficiaryResponse"}}}},"401":{"description":"UnAuthorized"}},"tags":["Beneficiary Registration APIs"],"security":[{"JWT-auth":[]}]},"get":{"operationId":"BeneficiaryController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Beneficiaries"}}}},"401":{"description":"UnAuthorized"}},"tags":["Beneficiary Registration APIs"],"security":[{"JWT-auth":[]}]}},"/beneficiary/{id}":{"get":{"operationId":"BeneficiaryController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Beneficiaries"}}}},"401":{"description":"UnAuthorized"},"404":{"description":"Could not found beneficiary"}},"tags":["Beneficiary Registration APIs"],"security":[{"JWT-auth":[]}]},"delete":{"operationId":"BeneficiaryController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Beneficiary removed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BeneficiaryResponse"}}}},"401":{"description":"UnAuthorized"},"404":{"description":"Could not found beneficiary"}},"tags":["Beneficiary Registration APIs"],"security":[{"JWT-auth":[]}]}},"/schedule":{"post":{"operationId":"ScheduleController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSchedule"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleResponse"}}}},"409":{"description":"Appointment already Scheduled"}},"tags":["Schedule Appointment APIs"]}},"/schedule/{id}":{"get":{"operationId":"ScheduleController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Schedule"}}}},"404":{"description":"Schedule data not found"}},"tags":["Schedule Appointment APIs"]},"patch":{"operationId":"ScheduleController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSchedule"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleResponse"}}}},"404":{"description":"Schedule data not found"}},"tags":["Schedule Appointment APIs"]},"delete":{"operationId":"ScheduleController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleResponse"}}}},"404":{"description":"Schedule data not found"}},"tags":["Schedule Appointment APIs"]}}}}